extract validation function from useItem.js
